requirements/dev.txt
-r base.txt
django-debug-toolbar==1.4
# Required by django-debug-toolbar
sqlparse==0.1.18

coverage==3.7.1
flake8==2.4.1
# Required by flake8
pyflakes==0.8.1
pep8==1.5.7
mccabe==0.3.1

requires.io
